requirements.txt
python-binance>=1.0.19
typer>=0.9.0
rich>=13.7.0
charset-normalizer==3.3.2
requests==2.31.0
cryptography>=41.0.0
